refactor: UserInfo를 valueObject로 분리
