+++ src/main/java/com/wanted/preonboarding/reservation/domain/entity/Reservation.java
@@ -4,14 +4,12 @@ import com.wanted.preonboarding.common.model.DefaultEntity;
 import com.wanted.preonboarding.common.model.SeatInfo;
 import com.wanted.preonboarding.performance.domain.entity.Performance;
 import com.wanted.preonboarding.reservation.domain.dto.ReservationRequest;
-import com.wanted.preonboarding.reservation.domain.dto.UserInfo;
+import com.wanted.preonboarding.reservation.domain.valueObject.UserInfo;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.util.UUID;
 
 @Entity
 @Table
